+++ wiki/ReleaseNotes.md
@@ -14,6 +14,9 @@
   reuses the upper and lower case as defined in the Franca model where possible
 * **[Java]** Add copy constructor to complex types of Franca model: for each complex data structure
   in the Franca model, a copy constructor is created in the respective Java class
+* **[Java, C++]** Rename subscription listener methods
+  * onReceive: Gets called on every received publication
+  * onError: Gets called on every error that is detected on the subscription
 
 ## Other changes
 * **[Tooling]** Enable cleanup capability of joynr generator framework: it is now possible to